// src/app/interfaces/car-product.interface.ts
export interface CarProduct{
    precio:       number;
    //id_producto:  string;
    cantidad?:    number;
    total?:       number;
    nombre?:      string;
    img?:         string;
    descripcion?: string;
    id?:          string;
    //precioFinal?: number;
    cantXproduct?: number;
    selectCant?: number;



}